pyproject.toml
# Kaggle Competition - 共通設定
# 各サブプロジェクトはこの設定を継承できます

[project]
name = "kaggle-competition"
version = "0.1.0"
description = "Kaggle competition project"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "torchvision",
    "pandas",
    "numpy",
    "scikit-learn",
    "pillow",
    "tqdm",
    "pyyaml",
    "timm",
    "trackio",
]

[project.optional-dependencies]
dev = [
    "jupyter",
    "pytest",
    "ruff",
    "pre-commit",
]

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.ipynb",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # Pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (formatterに任せる)
    "E731",   # lambda assignment (MLでよく使う)
    "B008",   # function call in default argument (PyTorchでよく使う)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
"notebooks/*" = ["E402", "F401", "B018"]  # import順序、未使用import、式文を許可
"**/experiments/*" = ["E402", "F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ty]
python-version = "3.10"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
